Fix remaining compiler warnings: sign conversion and format string
- digest.c: Cast to ssize_t instead of size_t when assigning negative values to bytes_skip_following_chunks
- appimagetool_sign.c: Use %lld format specifier with (long long) cast for off_t type

These fixes address all remaining -Werror warnings on all architectures.

+++ src/appimagetool_sign.c
@@ -213,7 +213,7 @@ bool embed_data_in_elf_section(const char* filename, const char* elf_section, gp
     }
 
     if (verbose) {
-        fprintf(stderr, "[sign] data size: %lu\n", data_size);
+        fprintf(stderr, "[sign] data size: %lld\n", (long long)data_size);
     }
 
     // rewind so we can later read the data
